restructures proxies, removes layer colors, and adds color proxy class
