+++ src/Speckle.Core/Models/Collections/Collection.cs
@@ -17,11 +17,9 @@ public class Collection : Base
   /// Constructor for a basic collection.
   /// </summary>
   /// <param name="name">The human-readable name of this collection</param>
-  /// <param name="collectionType"></param>
-  public Collection(string name, string collectionType)
+  public Collection(string name)
   {
     this.name = name;
-    this.collectionType = collectionType;
   }
 
   /// <summary>
