+++ src/Speckle.Core/Models/Collections/Layer.cs
@@ -3,7 +3,7 @@ namespace Speckle.Core.Models.Collections;
 /// <summary>
 /// A specialized collection that represents a CAD-app layer. We expect this to grow in the future with possibly other shared props.
 /// </summary>
-public class Layer : Collection, IHasColor
+public class Layer : Collection
 {
   public Layer() { }
 
@@ -11,15 +11,8 @@ public class Layer : Collection, IHasColor
   /// Constructor for a basic Layer.
   /// </summary>
   /// <param name="name">The human-readable name of this collection</param>
-  /// <param name="color"></param>
-  public Layer(string name, int color)
+  public Layer(string name)
   {
     this.name = name;
-    this.color = color;
   }
-
-  /// <summary>
-  /// The argb int value of the layer color
-  /// </summary>
-  public int color { get; set; }
 }
